+++ main/include/hal/drivers/display_st7701s.h
@@ -15,7 +15,7 @@ class display_st7701s : public Display {
   ~display_st7701s() override = default;
 
   esp_lcd_panel_handle_t getPanelHandle() override;
-  std::pair<int, int> getResolution() override { return {480, 480}; };
+  std::pair<int16_t, int16_t> getResolution() override { return {480, 480}; };
   bool onColorTransDone(esp_lcd_panel_io_color_trans_done_cb_t, void *) override;
   uint8_t *getBuffer() override;
   uint8_t *getBuffer2() override;
